jetson/requirements.txt
numpy>=1.19.4
adafruit-circuitpython-servokit>=1.3.0
adafruit-circuitpython-pca9685>=3.3.4
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
RPi.GPIO>=0.7.0
smbus2>=0.4.1
websockets>=10.0
psutil>=5.8.0
logging>=0.5.1.2
PyYAML>=5.4.1
